Program 4
Many students learn state capitals in the 5th or 6th grade. For this assignment, write a Python program that creates a state capital guessing quiz. The program prompts the user to enter the number of questions for the quiz. The program then generates a random selection of states and presents these states, one at a time, to the user. The user is to guess the state capital by entering the name from the keyboard. Your program should handle upper- or lower-case letters. After each guess, the program will display whether the user’s answer is correct or incorrect. If incorrect, the program will state the correct answer. Also, the program will keep track of how many answers are correct out of the total answered after each question and print the number correct after each question and at the end of the quiz. Finally, the program will output a bar chart that contains the number of correct and incorrect answers.

For this assignment, you must use a dictionary in which the keys are the state names, and the values are the corresponding state capitals. Here is a sample output from the program: